Fix hooks and add new matrix view

+++ src/org/powerbot/script/rt6/GroundItems.java
@@ -41,15 +41,15 @@ public class GroundItems extends GroundItemQuery<GroundItem> {
 		}
 		final int bx = base.x();
 		final int by = base.y();
-		for (long x = bx; x < bx + 104; x++) {
-			for (long y = by; y < by + 104; y++) {
+		for (int x = bx; x < bx + 104; x++) {
+			for (int y = by; y < by + 104; y++) {
 				id = x | y << 14 | plane << 28;
 				cache = org.powerbot.bot.rt6.HashTable.lookup(table, id, NodeListCache.class);
 				if (cache.isNull()) {
 					continue;
 				}
 				for (final ItemNode item : NodeQueue.get(cache.getDeque(), ItemNode.class)) {
-					items.add(new GroundItem(ctx, new Tile((int) x, (int) y, plane), item));
+					items.add(new GroundItem(ctx, new Tile(x, y, plane), item));
 				}
 			}
 		}
